% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children.R
\name{children}
\alias{children}
\alias{childrenof}
\title{List the children of something in a list tree}
\usage{
children(whose, where, key = "stId")

childrenof(obj)
}
\arguments{
\item{whose}{string. an idendifier indicating \code{key}}

\item{where}{list. the list tree including \code{whose}}

\item{key}{string. name we search depends on}

\item{obj}{list. the father node}
}
\description{
In a tree(e.g. a hierarchical structure for some species) represtented
by a nested list. \code{children} lists the children of \code{whose} as
a dataframe.
}
