% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_backup.R, R/reactomeDatabase.R
\name{reactomeDatabase}
\alias{reactomeDatabase}
\alias{rtDatabase}
\alias{reactomeDatabase}
\title{Database info queries}
\usage{
reactomeDatabase()

rtDatabase()

reactomeDatabase()
}
\value{
a dataframe.

A dataframe.
}
\description{
Return the name and version number of current database.

Return the name and the version of current database
}
\seealso{
\href{http://www.reactome.org}{\code{Reactome}}
}
