% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactomeDatabase.R, R/reactomeDiseases.R
\name{reactomeDatabase}
\alias{reactomeDatabase}
\alias{reactomeDiseases}
\title{Database Info Quenries}
\usage{
reactomeDatabase()

reactomeDiseases("diseases")
or rectomeDiseases("diseases/doid")
}
\arguments{
\item{dataType}{a chacter string that specify the data type. can be \code{diseasw}
or \code {disease/doid}}
}
\value{
the name and the version of current database

a dataframe
}
\description{
Return the name and the version of current database

retrieves the dataframe of disease DOIDs annotated in Reactome
}
\details{
Its Response Class (Status 200) is a string,else if Response Messages HTTP Status
\Code{406},the reason is Not acceptable according to the accept headers sent in the request
\code{500},Internal Server Error

It retrieves the dataframe of disease or disease DOIDs annotated in Reactome

'dataType = disease ' It retrieves the dataframe of diseases annotated in Reactome
'dataType = disease/doid',It retrieves the dataframe of disease DOIDs annotated in Reactome
Its Response Class (Status 200) is a string,else if Response Messages HTTP Status
\Code{406},the reason is Not acceptable according to the accept headers sent in the request
\code{500},Internal Server Error
}
\examples{
reactomeDiseases("diseases")
reactomeDiseases("diseases/doid")

}
