% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactomeEntity.R
\name{reactomeEntity}
\alias{reactomeEntity}
\alias{rtEntity}
\title{PhysicalEntity Queries}
\usage{
reactomeEntity(id, queryType, silent = FALSE)

rtEntity(id, queryType)
}
\arguments{
\item{id}{a character string. identifier of the complex, entity or the PhysicalEntity
that you want to query}

\item{queryType}{a character string that specify the query type. can be \code{subunit},
\code{structure}, or \code{form}.}
}
\value{
a dataframe
}
\description{
Retrieves the subunits, structures or other forms of a given entity
}
\details{
queryType = "subunits". Retrieves the list of subunits that constitute any given complex.

queryType = "componentOf". Retrieves the list of structures (Complexes and Sets) that include
the given entity as their component.

queryType = "otherForms". Retrieves a list containing all other forms of the given PhysicalEntity.
}
\examples{
reactomeEntity("R-HSA-5674003", "subunits")
reactomeEntity("R-HSA-199420", "componentOf")

entity = rtEntity("R-HSA-199420", "otherForms")

}
