% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children.R, R/eventHierarchy.R
\name{children}
\alias{children}
\alias{reactomeEventHierarchy}
\alias{rtEventHierarchy}
\title{List the children of something in a dataframe tree}
\usage{
children(whose, where, key = "stId")

reactomeEventHierarchy(species = "9606", silent = FALSE)

rtEventHierarchy(species = "9606", silent = FALSE)
}
\arguments{
\item{whose}{string. an idendifier indicating \code{key}}

\item{where}{list. the list tree including \code{whose}}

\item{key}{string. name we search depends on}

\item{species}{string. Allowed species filter: SpeciesName (eg: Homo sapiens)
SpeciesTaxId (eg: 9606)}

\item{silent}{logical. Whether print error messages on the screen.}
}
\value{
A data frame.
}
\description{
In a tree(e.g. a hierarchical structure for some species) represtented
by a nested dataframe. \code{children} lists the children of \code{whose}.

Events (pathways and reactions) in Reactome are organised in a hierarchical structure
for every species. By following all 'hasEvent' relationships, this method retrieves
the full event hierarchy for any given species. The result is a list of tree structures,
one for each TopLevelPathway. Every event in these trees is represented by a
PathwayBrowserNode. The latter contains the stable identifier, the name, the species,
the url, the type, and the diagram of the particular event.
}
\seealso{
\code{\link{reactomeEvent}}
}
