% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactor.R
\name{reactomeInteractor}
\alias{reactomeInteractor}
\alias{PSICQUIC_Provier}
\title{Molecule interactors}
\usage{
reactomeInteractor(acc, resource = 'Reactome', type = 'psicquic',
                   detail = TRUE, silent = FALSE)
reactomeInteractor(acc, type = 'static', detail = TRUE, page = -1,
                   pageSize = -1, silent = FALSE)
reactomeInteractor(acc, type = 'static', detail = FALSE, silent = FALSE)

PSICQUIC_Provier(silent = FALSE)
}
\arguments{
\item{acc}{string. Interactor accession(s) (or identifier(s)).}

\item{resource}{string. PSICQUIC Resource.}

\item{type}{string. \code{psicquic}: PSICQUIC method;
\code{static}: static method.}

\item{detail}{logical. Return details or summary.}

\item{page}{interger. For paginating the results.}

\item{pageSize}{interge. Number of results to be retrieved.}

\item{silent}{logical. Run quitely.}
}
\value{
\itemize{
  \item \code{logical FALSE}: When an error happens.
  \item \code{list}: When query details of multiple entries.
  \item \code{dataframe}: Others.
}
}
\description{
Retrieve clustered interaction, details or summary, sorted by score,
of single or multiple given accession(s) by resource i.e. PSICQUIC method.
Retrieve a detailed(summary) interaction information of given accession(s)
i.e. static method.
}
\details{
\code{PSICQUIC_Provier} returns the list of PSICQUIC Proviers.
}
\examples{
interactors = reactomeInteractor(acc = 'Q13501', resource = 'MINT', type = 'psicquic')
interactors = reactomeInteractor(acc = 'Q13501', resource = 'MINT', type = 'static')

interactors = reactomeInteractor(acc = c('Q13501','Q16543'),
                                 resource = 'MINT', type = 'psiquic')
interactors = reactomeInteractor(acc = c('Q13501','Q16543'),
                                 resource = 'MINT', type = 'static')
}
