% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowerPath.R
\name{rtLowerPathway}
\alias{rtLowerPathway}
\alias{rtTopPathway}
\title{Pathway related queries}
\usage{
rtLowerPathway(id, species = "48887", allForms = TRUE, withDiagram = TRUE,
  silent = FALSE)

rtTopPathway(species, silent = FALSE)
}
\arguments{
\item{id}{string. The entity (in any of its forms) or a identifier that has to be present
in the pathways.}

\item{species}{string.The species for which the pathways are requested
(SpeciesName or SpeciesTaxId)}

\item{allForms}{logical;if TRUE, any form of the given PhysicalEntity or
Event is consideried. default = TRUE}

\item{withDiagram}{logical;if TRUE, only those pathways with a diagram will be listed.
default = FALSE.}

\item{silent}{logical; if TRUE, error message will not be printed. default = fasle}
}
\value{
a data frame
}
\description{
The \code{rtLowerPathway} function traverses the event hierarchy and retrieves the list of
all lower level pathways that have a diagram(optional) and contain any form
(optional) of the given PhysicalEntity or Event.
The \code{rtTopPathway} function retrieves the list of top level pathways for the given species.
}
\examples{
res.1 = rtLowerPathway(id = "R-HSA-199420", species = "48887",
                  allForms = TRUE, withDiagram = TRUE)
# all those pathways with or without a diagram
res.2 =  rtLowerPathway(id = "R-HSA-199420", species = "48887",
                  allForms = TRUE, withDiagram = FALSE)
# search with a known  identifier
res.3 =  rtLowerPathway(id = "PTEN", species = "48887",
                  allForms = TRUE, withDiagram = TRUE)
# find all top level pathways for Homo sapiens
human.1 = rtTopPathway(species = "9606")
# SpeciesName is also permitted for rtTopPath
human.2 = rtTopPathway(species = "Homo sapiens")
}
