% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people.R
\name{people}
\alias{people}
\title{Person queries}
\usage{
people(id, type = NULL)
}
\arguments{
\item{id}{string.Person identifier: Can be OrcidId, DbId. Email can also be used for
type = \code{authoredPathways} or \code{publications}}

\item{type}{string. Can be NULL, a attribute to be filtered, \code{authoredPathways} and
\code{publications}.}
}
\value{
a data frame.
}
\description{
Retrieves a specific person's property by his/her OrcidId or DbId or a list of pathways or publications
authored by a given person. OrcidId, DbId or Email can be used to specify the person.
}
\details{
When type = NULL, all attributes about the given person in Reactome will be retrieved unless a single
attribute name is specified by the "type" parameter.

A list of pathways or publications authored by the given person will be returned with
type = \code{authoredPathways} or \code{publications}.
}
\examples{

# Retrieves a person in Reactome whose orcidId is 0000-0001-5807-0069
res = people(id = "0000-0001-5807-0069", type = NULL)


##
res = people(id = "0000-0001-5807-0069", type ="authoredPathways" )

##
res = people(id = "0000-0001-5807-0069", type = "publications")
}
