% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person.R
\name{rtperson}
\alias{rtperson}
\alias{authoredPathway}
\alias{authoredPublication}
\title{Person queries}
\usage{
rtperson(id, attributeNames, ...)
rtperson(name, exactlyMatch, ...)

authoredPathway(id, silent = FALSE)

authoredPublication(id, silent = FALSE)
}
\arguments{
\item{id}{string or charactor vector.Person identifier: Can be OrcidId, DbId. Email can also be used for
the methods \code{authoredPathway()} or \code{authoredPublication()}}

\item{name}{string or character vector.first  or last name of people you want to search in Reactome.}

\item{exactlyMatch}{logical.If TRUE, the default, a list of people in Reactome with either their first
or last name matching exactly the given string(\code{name}) will be returned for name query method.}

\item{attributeNames}{string. Parameter for id query method. Attribute names to be filtered.
If NULL, the default, no filtration wil be done.}

\item{silent}{logical. Run quitely.}
}
\value{
a data frame.
}
\description{
\code{rtperson()} retrieves a specific person's information in Ractome by his/her OrcidId or DbId.
\code{authoredPathways()} or \code{authoredPublication()} can be used to get a list of
 pathways or publications authored by a given person.
}
\details{
When exactlyMatch = false, \code{name} should be a string or character vector of length 1.
Only one of \code{id} and \code{name} needs to be specified, the other remains null.
}
\examples{
id.query = rtperson(id = c("0000-0001-5807-0069","5251225"),
                   attributeNames = c("displayName", "firstname"), silent = FALSE)
# id.query = rtperson(id = c("0000-0001-5807-0069","5251225"),
#                  attributeNames = NULL, silent = TRUE)
# search by name
name.query = rtperson(name = c("Steve Jupe", "Tam Angela"), exactlyMatch = TRUE, silent = FALSE)

# get a list of pathways or publications authored by given people
dt.1 = authoredPublication(id = "0000-0001-5807-0069",silent = FALSE)
dt.2 = authoredPathway(id = "0000-0001-5807-0069",silent = FALSE)
}
