% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtReferences.R
\name{reactomeReferences}
\alias{reactomeReferences}
\alias{rtReferences}
\title{ReferenceEntity queries}
\usage{
reactomeReferences(id)

rtReferences(id)
}
\arguments{
\item{id}{string. Identifier for a given entity}
}
\value{
A dataframe.
}
\description{
Retrieves a dataframe containing all the reference entities for a given identifier.
}
\examples{
refrenceEntity = reactomeReferences("15377").

}
