% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtContainedEvents.R
\name{rtContainedEvents}
\alias{rtContainedEvents}
\title{Pathway related queries}
\usage{
rtContainedEvents(id, ATTR = "all", silent = TRUE)
}
\arguments{
\item{id}{string. The event for which the contained events are requested.}

\item{ATTR}{string. Retrieve all attributes or only a single attribute for each of the
events contained in the given event. default = \code{all}}

\item{silent}{logical; if TRUE, error message is not printed.}
}
\value{
a data frame.
}
\description{
This method recursively retrieves all the events contained in any given event.
}
\details{
Events are the building blocks used in Reactome to represent all biological processes,
and they include pathways and reactions. Typically, an event can contain other events.
For example, a pathway can contain smaller pathways and reactions.

This function will get all infromations about the containes events when ATTR = \code{"all},
otherwise, only the attribute sepcified by \code{ATTR} will be retrieved.
}
