% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowerPath.R
\name{rtLowerPath}
\alias{rtLowerPath}
\title{Pathway related queries}
\usage{
rtLowerPath(id, species = "48887", type = "entity", allForms = TRUE,
  diagram = TRUE, silent = TRUE)
}
\arguments{
\item{id}{string. The entity (in any of its forms) that has to be present
in the pathways.}

\item{species}{string.The species for which the pathways are requested
(SpeciesName or SpeciesTaxId)}

\item{type}{string. Whether the given \code{id} is a entity or identifier.}

\item{allForms}{logical;if TRUE, any form of the given PhysicalEntity or
Event is consideried. default = TRUE}

\item{diagram}{logical;if TRUE, only those pathways with a diagram will be listed.
default = TRUE.}

\item{silent}{logical; if TRUE, error message will not be printed. default = TRUE}
}
\value{
a data frame
}
\description{
This method traverses the event hierarchy and retrieves the list of
all lower level pathways that have a diagram(optional) and contain any form
(optional) of the given PhysicalEntity or Event.
}
\examples{
res = rtLowerPath(id = "R-HSA-199420", species = "48887", type = "entity",
                  allForms = TRUE, diagram = TRUE)

res =  rtLowerPath(id = "R-HSA-199420", species = "48887", type = "entity",
                  allForms = TRUE, diagram = FALSE)

res =  rtLowerPath(id = "PTEN", species = "48887", type = "identifier",
                  allForms = TRUE, diagram = TRUE)
}
