% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopPath.R
\name{rtTopPath}
\alias{rtTopPath}
\title{Pathway related queries}
\usage{
rtTopPath(species)
}
\arguments{
\item{species}{string. Specifies the species by SpeciesName (eg: Homo sapiens)
or SpeciesTaxId (eg: 9606).}
}
\value{
a data frame
}
\description{
This method retrieves  the list of top level pathways for the given species.
}
\examples{
res = rtTopPath(species = "9606")
}
