% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{reactomeDatabase}
\alias{reactomeDatabase}
\alias{rtDatabase}
\title{Database info queries}
\usage{
reactomeDatabase()

rtDatabase()
}
\value{
a dataframe.
}
\description{
Return the name and version number of current database.
}
\seealso{
\href{http://www.reactome.org}{\code{Reactome}}
}
