% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactomeDiseases.R
\name{reactomeDiseases}
\alias{reactomeDiseases}
\title{Disease related queries}
\usage{
reactomeDiseases(DOID_only = FALSE)
}
\arguments{
\item{DOID_only}{logical. \code{FALSE}: return DOIDs of diseases only.
\code{TRUE}: return detailed infomation.}
}
\value{
a dataframe
}
\description{
It retrieves the dataframe of disease or disease DOIDs annotated in Reactome
}
\examples{
diseases = reactomeDiseases()
diseases_doid = reactomeDiseases(TRUE)

}
