% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactomeEntity.R
\name{rtEntity}
\alias{rtEntity}
\title{PhysicalEntity Queries}
\usage{
rtEntity(id, queryType, simplify = TRUE)
}
\arguments{
\item{id}{a character vector of identifiers of the complex, entity or the PhysicalEntity
that you want to query}

\item{queryType}{a character vector that specify the query type. can be \code{subunits},
\code{componentOf}, or \code{otherForms}.}

\item{simplify}{attempt to reduce the result to a data frame when length of \code{id} is 1.}
}
\value{
a list of tha same length of \code{id}, with ecah element corresponding to the query result of
one of id.
}
\description{
Retrieves the subunits, structures or other forms of a given entity
}
\details{
queryType = \code{subunits}. Retrieves the list of subunits that constitute any given complex.
In case the complex comprises other complexes, this method recursively breaks all of them into
their subunits.

When queryType = \code{componentOf}, this function will retrieves the list of structures
(Complexes and Sets) that include the given entity as their component.It should be mentioned
that the list includes only simplified entries (type, names, ids) and not full information about each item.

queryType = \code{otherForms}. Retrieves a list containing all other forms of the given PhysicalEntity.
These other forms are PhysicalEntities that share the same ReferenceEntity identifier,
e.g. PTEN H93R[R-HSA-2318524] and PTEN C124R[R-HSA-2317439] are two forms of PTEN.
}
\examples{
rtEntity("R-HSA-5674003", "subunits")
rtEntity("R-HSA-199420", "componentOf")
rtEntity(id = c("R-HSA-5674003", "R-HSA-199420"),
         queryType= c("subunits", "componentOf"))

}
