% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier.R
\name{rtAnalyseIdentifier}
\alias{rtAnalyseIdentifier}
\title{Queries for one or multiple  identifiers}
\usage{
rtAnalyseIdentifier(id, ...)

rtAnalyseIdentifier(file, ...)
}
\arguments{
\item{id}{character vector. one or multiple identifiers to analyse followed by
their expresssion(when applied)}

\item{file}{path to file that contains data to be analysed.}

\item{MakeProjection}{logical, whether projects the result to Homo Sapiens or not.}

\item{IncludeInteractors}{logical. whether include interactors.}

\item{pageSize}{integer. pathways per page.}

\item{page}{integer. page number}

\item{sortBy}{string. how to sort the result.}

\item{order}{string. specifies the order.}

\item{resource}{string. the resource to sort.}

\item{silent}{logical, to run quietly.}
}
\value{
a list with a data frame including pathway information and the token
object; if no pathway found, only a summary data frame will be returned.

The summary.token can be used to retrieve the results of a previously
performed analysis without the need to submit the sample again.
}
\description{
Analyse the post identifiers over the different species or projects the result
to Homo Sapiens
}
\details{
The projection is calculated by the orthologous slot in the Reactome database.
Use page and pageSize to reduce the amount of data retrieved. Use sortBy and
order to sort the result by your preferred option. The resource field will
filter the results to show only those corresponding to the preferred molecule
type (TOTAL includes all the different molecules type).

\code{sortBy} can be one of \code{NAME,TOTAL_ENTITIES,TOTAL_INTERACTORS,
TOTAL_REACTIONS,FOUND_ENTITIES,FOUND_INTERACTORS,FOUND_REACTIONS,ENTITIES_RATIO,
ENTITIES_PVALUE,ENTITIES_FDR,REACTIONS_RATIO}.

\code{order} can be \code{ASC} or \code{DESC}.

\code{resource} can be \code{TOTAL,UNIPROT,ENSEMBL,CHEBI,MIRBASE,NCBI_PROTEIN,
EMBL,COMPOUND}.
}
\examples{
# analyse multiple ids at the same time
pathwayInf.1 = rtAnalyseIdentifier(id = c('PTEN', 'PIK3C2A', 'UNC58'))
# projects the result to Homo Sapiens
pathwayInf.2 = rtAnalyseIdentifier(id = c('PTEN', 'PIK3C2A', 'UNC58'),
MakeProjection = TRUE)
}
